[Disco] Dynamic-shape parameter loading in TP (#1632)
This PR fixes the issue introduced due to the dynamic-shape
parameter support.

The way is to introduce a temporary data structure called
`SymShapeTuple` in model metadata reading. The `SymShapeTuple`
instance may contain `-1` in some of its dimensions, and have
a corresponding symbolic name on such dimensions. The symbolic
shape tuple will be instantiated to `ShapeTuple` with
the concrete shape in model config.