[Serving] Stop string matching algo enhancement (#1711)
This PR enhances the stop string matching algorithm.

Prior to this PR, our stop string matching works on frontend
after UTF-8 decoding, and always caches a suffix string that
has length not exceeding the stop string length. This algorithm
is not optimal though, since it may delay the response time
for a request til the first token, due to the string suffix
cache.

This PR makes two changes:
1. applying KMP algorithm for linear stop string matching
with minimum caching,
2. move the stop string handling from front-end back into
the engine side, so that frontend only handles UTF-8.

The stop string is validated to be efficient, as tested in
`tests/python/support/test_streamer.py`.

----------------------

NOTE: this PR bumps `3rdparty/tokenizers-cpp`. You may need
to remove the CMakeCache.txt and build mlc-llm again. Or
otherwise the incremental build may fail, since the changes
in tokenizer Rust file doesn't seem to be compiled in
incremental build.