[Transform] Handle non-schedulable func in LiftTIRGlobalBufferAlloc
Prior to this commit, the implementation of
`mlc_llm.transform.LiftTIRGlobalBufferAlloc` assumed that every
PrimFunc in a module was schedulable.  If the module contained a
non-schedulable PrimFunc, the `assert isinstance(func.body,
tir.BlockRealize)` would produce an error.

This commit updates `LiftTIRGlobalBufferAlloc` to leave unrecognized
PrimFunc instances, and their callsites, unmodified.