[Serving] Make RequestState as a standalone object class
This PR adopts suggestions from the support of OpenAI API parallel
generation `n` in #1868. The main update in this PR is to make
the RequestState as a standalone object class, which was a typedef
from `std::vector<RequestStateEntry>` before.

This PR also fixes a bug in prefill that will cause engine failure
when `n` is large.