[Serving] Estimate KV cache memory usage with metadata
Prior to this PR, the serving engine memory usage estimation reads
model config for fields such as `num_key_value_heads`,
`num_hidden_layers`, etc.. However, since not every model share the
same set of config names (#1854), the estimation fails for models
that do not have this set of config field names.

This PR makes the following changes. First, it attaches these
field values into the model's metadata, in which way we unify the
field names for different models effectively. Then, when estimating
the memory usage, we read these fields from the metadata, rather than
model config, so we are safe for the name inconsistency.