[Serving] Address embedding lookup OOM issue
This PR addresses the OOM issue that may be caused by embedding
lookup when the batch size of a prefill action is large.
Prior to this PR, a large embedding tensor will be created for
each sequence in the prefilled batch, thus may take unexpectedly
large memory when the batch size is large.