[Model] Eliminate the reshape in embedding func
Prior to this PR, there is a trailing reshape kernel at the end of
the embedding func. The reshape is not necessarily needed to be
as a kernel, which consumes extra time during execution. This PR
eliminates the reshape in the embedding function by updating the
signature of the embedding func, so that now it only takes the plain
1D token ids as input.