[Serving] Avoid unnecessary worker sync in Model
Following up #1906, this PR removes the synchronization given it is
avoidable. We use another approach to avoid the write-after-write
issue.

The key to address the issue is to make sure the addresses to be
copied to worker 0 is not rewritten before the copy actually happens.
So we pre-allocate a large host array to hold all the token ids,
and for each sequence, we copy its token ids to the offset given
when calling TokenEmbed, so that we can make sure an address will
not be written twice before copy happens.