[Serve] Constrain KV cache capacity on Metal
This PR constrains the KV cache capacity for Metal devices to 32768,
in order to avoid large tensors in KV cache. This is because right
now Metal runtime has performance issue when running a kernel where
when some input buffer is very large, even if little of the large
buffer is accesed in the kernel.