[Serving][Fix] Fix JSON output check in test_server.py (#1966)
`test_server::is_json_or_json_prefix` is used to check the output
is JSON or a prefix of JSON. 

It uses json.loads internally. However, json.loads (i.e. json.decode)
is token-based instead of char based. If half a token is left at the
end of the string, it cannot be matched.

This PR adds another check for the rest "half a token" if it exists.