[Compiler] Support AUTO mode for all-reduce strategy (#2034)
This PR supports the auto mode for IPC all-reduce strategy.
It renames the strategy from `allreduce-strategy` to
`ipc-allreduce-strategy` in the compiler optimization flags. The
default RING mode is renamed to NONE mode, which, when specified,
uses nccl all-reduce without any IPC memory rewrite.

So right now to enable IPC all-reduce, the ideal way is to do
`ipc-allreduce-strategy=auto`.