[JIT] Better organize JIT and AOT handling
Previously we do JIT when AOT lib lookup failed.
The error message can become cryptic when JIT also fails,
it will show up as cannot find None-vulkan.dll.

This PR changes the behavior to only to lookup when model_lib_path
is provided, or only to JIT when it is not. This will leads to
cleaner error message overall.