[Serving] Support loading system library (#2189)
This PR introduces the support of loading system libraries. Now in
engine reload, when the given library path starts with `"system://"`,
we recognize this as a system library and will try to load the the
library from the path after the `"system://"` prefix.

This PR also decouples the InitBackgroundEngine of ThreadedEngine
into two parts, where the reload is now called explicitly when
initializing the engine. This can be also done for the JSONFFIEngine.
However, we need to move the construction of streamers in JSONFFIEngine
before doing the same thing for JSONFFIEngine. So this is marked as
a TODO item.