[Pass] Support two-stage softmax (#2220)
This PR introduces the compiler pass that rewrites the normal softmax
to a two-stage softmax. This is based on our finding that when
vocabulary size is large, the normal softmax cannot have high-enough
parallelism on GPU. So we partition the workload into two stages
for better parallelism and better performance.