[GenConfig] Set upper bound for prefill chunk size
By default the prefill chunk size is set to the context window size
or the sliding window size. When the number is large, our memory
planning during model compilation will allocate a lot memory.

Given we have support for input chunking, we can reduce the prefill
chunk size to a small value to save runtime memory.

This PR sets the prefill chunk size to be at most 2048.