[iOS] Initial scaffolding of MLCEngine in Swift (#2279)
[iOS] Initial scaffolding of LLMEngine in Swift

This PR adds initial scaffolding of LLMEngine in swift.
We wraps callback to AsyncStream so it can be accessed using for await API.

We also added an minimal example app to showcase the new MLCEngine,
the old ChatModule is still used in the MLCChat App.

The return value is structified already.
We will still need to structurify the chat completion interface.