[Serving] Fix BatchVerify to feed the extra token when fully accepted
This PR fixes a bug in the BatchVerify action.
When a draft model's proposal is fully accepted by the main model, there
is an extra token which is already in the main model's KV cache but not
in the draft model's KV cache.

Prior to this PR, BatchVerify action does not feed this extra token into
the draft model's KV cache, which causes size mismatch between the
main model's KV cache and draft model's KV cache.

This PR fixes this issue by adding an additional BatchDecode step for
the requests whose draft proposals are fully accepted by the main model.