[Serving][Refactor] Metrics and stats for CLI (#2387)
This PR introduces the `Metric` class for convenient metric update
and management in MLC. The previous `EngineStats` class is renamed
to `EngineMetrics` accordingly.

This PR brings the metric support to JSONFFIEngine, and implements
the `/stats` command in CLI.

Besides, this PR

* fixes a bug of time measurement when parallel generation exists.
* aligns the metric names with LLMPerf (particularly, we now use
`num_input_tokens`, `num_output_tokens`, `sum_num_input_tokens`, etc.)
* measures the time of a single step of BatchDecode, a single step
of draft generation in BatchDraft, and a single step of BatchVerify
when the effective batch size is less than 64 (hardcoded as a constant
as of now). This can help build the understanding of the performance
of the key actions under a series of batch size.