[Fix] Fix metrics division by 0
This PR fixes an issue of the per-request metrics, where division-by-0
may happen when the request does not run any decode step.

The division-by-0 results in `inf`, and is added into a JSON file.
However, `inf` is usually not recognized as a float value in JSON
grammar. Thus JSON parsers fail on parsing any JSON string that comes
with `inf` wihtout being quoted.