[Serving] Use stop strs and token ids for completions
This PR applies the stop strings and stop token ids defined in
conversation tempalte to the raw text completions. So that whenever
the model outputs a stop token id or stop string, the raw generation
can stop.

Prior to this commit, the raw text never stops when the max tokens
is not given. This commit helps reduce the frequency of such events.
Nevertheless, if the model does not output a stop string/token id,
the generation will still not be going to stop.