[Engine] Lazy recompute in GetRunningRequestStateEntries (#2655)
This PR updates GetRunningRequestStateEntries to make it lazy.
We use a dirty flag to check whether the running request state entries
are changed since the last recompute.

We make this improvement due to the observation that this function
may cause some CPU overhead. During consecutive rounds of batch decode,
the running requests don't change, so we can effectively use this
dirty flag to avoid recomputation.