[Bench] Adopting multi-processing to send requests
This PR adopts multi-processing to send requests in the benchmark
when the number of concurrent requests is larger than 32. This is
because when the number of concurrent requests is too high, asyncio
may have large pressure and won't be able to send requests fast on
time.

One caveat is that with multi-processing, we are not able to display
an accurate progress bar with tqdm as before.