[Bench] Support benchmarking for fixed request rates
This PR introduces the benchmark support for fixed request rates.
Specifically,

* We introduced the `AttachRequestRateTimestamp` request processor
which attaches timestamps according to the specified request rate
with regarding to the Poisson distribution.
* We introduced the `FixedConcurrentRequestExecutor` executor which
launches requests at the specified time according to the timestamps.
This executor can be used for both the fixed request rate case and
the log replay case (yet to be introduced).