[Python] Reuse KVCache classes from TVM
This PR updates the KV cache structuring to reuse the classes
from TVM.

This PR also includes the changes in #2847 which updates the KV cache
constructor regarding the RoPE extension factors. The intention is to
keep the attention interface unchanged, and it is more allowed to
update the KV cache constructor interface.