[Engine] Preparation for switching between spec-decode mode and normal mode
This PR introduces some prerequisite operations to support automatic
switching between speculative decoding mode and the normal batch decode
mode.

Specifically, this PR
* removes the constraint that the BatchDecode action is applicable
when there is only one model.
* enhances the BatchDraft action so whenever the committed tokens
of a request in the draft model is lagging behind the committed
tokens in the main model, the BatchDraft action will automatically
prefill all the lagging-behind part into the draft model when
generating the draft.

These changes are tested locally.