[Fix] Attribute mismatch in Phi-3-Vision Model (#2862)
This PR fixes the attribute mismatch in the Phi3 Vision model

Additionally, we are using the RoPE mode "long" for Phi3 Vision,
even though the RoPE mode in config.json is set to "su". This change
is due to the warning message: "The class Phi3SuScaledRotaryEmbedding
is deprecated and will be removed in version 5 of Transformers.
Please use Phi3LongRoPEScaledRotaryEmbedding instead," as mentioned
in modeling_phi3.py in the HF/Transformers library.