[Engine][Spec] Initiali adaptive speculative decoding
This PR introduces the initial support of speculative decoding,
which is enabled when the `spec_draft_length` is 0 (its default
value is also changed to 0 in this PR so that the adaptive mode
is enabled by default).

The goal of adaptive speculative decoding is to dynamically adjust
the draft length depending on the engine's running state. For the
initial version, we use a fixed table for the draft length selection.
We will follow up with more powerful draft length selection.

This PR also fixes a few bugs after the introduction of tree draft.