[Bench] Use OpenAI v1/completions as default (#2918)
This PR updates the benchmark to use OpenAI entrypoint `v1/completions`
as the default to remove the impact of `v1/chat/completions` on
conversation template and system prompt. The previous endpoint
is renamed to `openai-chat`, and the naming here is aligned with vLLM.

This PR also fixes the TPOT calculation. Previously it used "number of
output tokens - 1" as the divisor, and this PR changes "1" to the number
of tokens in the first output chunk.