[Bench] Fix the dataset truncation
This PR fixes the benchmark dataset processing. We used
`tokenizer.model_max_length` as the length of truncation, but for some
tokenizers (e.g., Llama3), the value of `model_max_length` is an
infinitely large integer, which causes error in the truncation.

This PR fixes it by setting an internal maximum limit of 100k.