[Bench] Specialize OpenAI endpoint for backends not supporting debug config
This PR specializes the OpenAI endpoint with the option of "not
attaching `debug_config`". Prior to this PR, when `--ignore-eos` is
enabled, a `debug_config` field is included in the payload for MLC
to recognize. However, some backend servers (e.g., vLLM) do not
support this `debug_config` field. So we need to manually specify
not attaching this field when benchmarking these backends.