[Bench] Add "per-gpu-workload" mode
This PR introduces the per-gpu-workload mode to MLC bench.
Under this mode, the specified "num_concurrent_requests" and
"request_rate" denote the workload **per GPU**, which means the overall
workload of the entire serving system for benchmarking will be
multiplied by the number of GPUs.

Meanwhile, this PR deprecates the argument `--testset-name` in
favor of `--dataset-path` for Loogle dataset.