[Model] Add support for Nemotron architecture (#3069)
This PR adds support for Nemotron architecture, and is in reference
to https://github.com/mlc-ai/mlc-llm/issues/2901 [Request for Nemotron-Mini-4B-Instruct]

Based on my analysis of the Nemotron architecture in
the huggingface repository, it appears to share similarities
with the Llama architecture, but with the following key distinctions:

- The activation function used in the MLP is `relu2` (squared ReLU).
- The MLP includes `up_proj` and `down_proj`, but does not have
a `gate_proj` as seen in Llama.
- It uses `layernorm1p`, and the normalization layer incorporates a bias term.
- The architecture employs a `partial_rotary_factor`, which is similar
to the approach used in the Phi architecture.