[Fix] Fix gemma 3 conv template stop token (#3191)
Currently for all the gemma3 instruction models, we reused the previous
gemma conv template. However, there is a difference in the stop token.

- In gemma2, the stop tokens are 1, 107: https://huggingface.co/google/gemma-2-2b-it/blob/main/generation_config.json
- In gemma3, the stop tokens are 1, 106: https://huggingface.co/google/gemma-3-1b-it/blob/main/generation_config.json

This PR add a new `gemma3_instruction` conv template, and will manually
update the already-uploaded gemma 3 models on HF.

In gemma3, 107 is `\n`. So currently whenever the model tries to change to
a new line, it ends generation.