[MultiGPU] Fix dist runtime device specification
This PR fixes a bug of device specification. Prior to recent
runtime refactors, the `Device null_device{DLDeviceType(0), 0}`
is treated as a null for `Optional<Device>`, but this is no
longer true in the latest runtime. Therefore, we use
`Optional<Device>(std::nullopt)` to replace the previous `null_device`
for a correct device specification.