[FFI] Rename `NDArray` to `Tensor` (#3326)
This PR applies the recent TVM ffi refactor that renames NDArray
to Tensor, to align with the PyTorch name convention.

Aside from the rename, this PR also updates the device API renaming
and the `register_global_func` renaming.

---

NOTE: This PR is a breaking changes:

- all compiled model libraries are invalidated. Please recompile all
models with `mlc_llm compile` or `MLC_JIT_POLICY=REDO`.

- `ndarray-cache.json` is now renamed to `tensor-cache.json`. Please
apply this rename in your local model directories.