[C++] Fix EventTraceRecorder initialization (#3331)
This PR fixes the initialization of EventTraceRecorder after the
enhancement of ObjectRef null safety.

Prior to this PR, EventTraceRecorder is designed as a not-nullable
class, while it doesn't have a corresponding constructor and is
initialized by `make_object` in an external function. This is
considered invalid with the ObjectRef null safety enhancement.
Therefore, we relax the requirement and allow it to be nullable.